+++ rubrics-ui/src/app/rate-rubric/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useState, useEffect } from 'react';
-import RubricsPage from '@/components/rubricsTable/rubricsTable';
+import RubricsPage from '@/components/RubricsView/RubricsPage';
 import { RubricServerData } from '@/types/rubricsTypes/types';
 import { RateRubricProps } from '@/types/rubricsTypes/types';
 import { getSession } from 'next-auth/react';
