+++ rubrics-ui/src/components/RubricsView/RubricDetails.tsx
